refactor: move schedule-related functions to separate file from utils
