+++ src/components/Calendar.tsx
@@ -1,13 +1,14 @@
 "use client";
 import { Class } from "@/lib/definitions";
 import { ColorsEnum, DaysEnum } from "@/lib/enums";
-import { cn, getCardColors } from "@/lib/utils";
-import { useCallback, useState } from "react";
+import { cn, formatTime, getCardColors } from "@/lib/utils";
+import { useCallback, useEffect, useState } from "react";
 import CalendarCard from "./CalendarCard";
 import { ScrollArea } from "./ui/scroll-area";
 
 const CELL_SIZE_PX = 64;
 const CELL_HEIGHT = "h-16";
+const TOP_OFFSET = 16; // Based on 16px (1rem) padding in the calendar
 
 const Calendar = ({
   courses,
@@ -23,7 +24,10 @@ const Calendar = ({
   isMobile?: boolean;
 }) => {
   const calculateHeight = useCallback(
-    (start: number, end: number) => {
+    (start: number, end: number, type: "military" | "minutes" = "military") => {
+      if (type === "minutes") {
+        start;
+      }
       const startHour = Math.floor(start / 100);
       const endHour = Math.floor(end / 100);
       const startMinutes = start % 100;
@@ -39,6 +43,12 @@ const Calendar = ({
   );
 
   const [hovered, setHovered] = useState<number | false>(false);
+  const [dragging, setDragging] = useState(false);
+  const [selection, setSelection] = useState<{
+    start: number;
+    end: number;
+    day: DaysEnum;
+  } | null>(null);
 
   const sortedClasses = courses.reduce<
     Record<DaysEnum, (Class & { color: string; shadow: string })[]>
@@ -63,6 +73,77 @@ const Calendar = ({
   const headerStyle =
     "relative h-full w-full text-center py-2 px-2 mx-2 font-bold text-muted-foreground";
 
+  const handleMouseDown = (e: React.MouseEvent) => {
+    const LEFT_OFFSET = 66; // Based on 50px + 1rem (16px)
+
+    const xPos = e.clientX - e.currentTarget.getBoundingClientRect().left;
+    const yPos = e.clientY - e.currentTarget.getBoundingClientRect().top;
+
+    if (xPos < LEFT_OFFSET || yPos < TOP_OFFSET) return; // Ignore clicks in the time column or extra space
+
+    const sizePerColumn = Math.floor(
+      (e.currentTarget.getBoundingClientRect().width - LEFT_OFFSET) / 6
+    );
+
+    const column = Math.floor((xPos - LEFT_OFFSET) / sizePerColumn);
+    const timeSlot = Math.floor((yPos - TOP_OFFSET) / (cellSizePx / 4));
+
+    const startTime = 7 * 60 + 15 * timeSlot; // 7:00 AM + 15 minutes per slot
+
+    setDragging(true);
+    setSelection({
+      start: startTime,
+      end: startTime,
+      day: ["M", "T", "W", "H", "F", "S"][column] as DaysEnum,
+    });
+
+    console.log(formatTime(startTime, "minutes"));
+  };
+
+  const handleMouseUp = (e: MouseEvent) => {
+    setDragging(false);
+  };
+
+  const handleMove = (e: React.MouseEvent) => {
+    if (!dragging || !selection) return;
+
+    const yPos = e.clientY - e.currentTarget.getBoundingClientRect().top;
+
+    if (yPos < TOP_OFFSET) return; // Ignore clicks in the time column or extra space
+
+    const timeSlot = Math.floor((yPos - TOP_OFFSET) / (cellSizePx / 4));
+
+    const rawEnd = 7 * 60 + 15 * timeSlot; // 7:00 AM + 15 minutes per slot
+
+    // If the card is going down, we need to make it add +15 minutes since that's the minimum time slot
+    const slotOffset = selection.start <= rawEnd ? 15 : 0;
+
+    const endTime = rawEnd + slotOffset;
+
+    setSelection({
+      ...selection,
+      end: endTime,
+    });
+    console.log(
+      formatTime(selection!.start, "minutes"),
+      formatTime(endTime, "minutes")
+    );
+  };
+
+  useEffect(() => {
+    if (window === undefined) return;
+
+    if (dragging) {
+      window.addEventListener("mouseup", handleMouseUp);
+    } else {
+      window.removeEventListener("mouseup", handleMouseUp);
+    }
+
+    return () => {
+      window.removeEventListener("mouseup", handleMouseUp);
+    };
+  }, [dragging]);
+
   return (
     <div className="flex flex-shrink min-h-0 w-full flex-col border rounded-lg bg-background">
       {/* Day Indicator Row */}
@@ -81,7 +162,11 @@ const Calendar = ({
       {/* Scrollable Container */}
       <ScrollArea>
         {/* Calendar Content */}
-        <div className="flex h-max w-full flex-row">
+        <div
+          className="flex h-max w-full flex-row"
+          onMouseDown={handleMouseDown}
+          onMouseMove={handleMove}
+        >
           {/* Time Column */}
           <div className="ml-2 flex w-[50px] shrink-0 flex-col items-end">
             {[...Array(16)].map((_, index) => (
@@ -122,6 +207,17 @@ const Calendar = ({
                   }`}
                   key={day}
                 >
+                  {selection?.day === day && (
+                    <div
+                      style={{
+                        height: calculateHeight(selection.start, selection.end),
+                        top: calculateHeight(700, selection.start) + TOP_OFFSET,
+                      }}
+                      className="bg-primary"
+                    >
+                      Test
+                    </div>
+                  )}
                   {sortedClasses[day].map((currClass) => {
                     const schedules = currClass.schedules.filter(
                       (sched) => sched.day === day
@@ -137,7 +233,7 @@ const Calendar = ({
                           currClass={currClass}
                           sched={sched}
                           height={calculateHeight(start, end)}
-                          top={calculateHeight(700, start) + 16}
+                          top={calculateHeight(700, start) + TOP_OFFSET}
                           hovered={hovered}
                           onMouseEnter={() => setHovered(currClass.code)}
                           onMouseLeave={() => setHovered(false)}
